Added percolation on USTs + Cluster Analysis functionality

+++ src/wilson.c
@@ -21,29 +21,29 @@ uint8_t init_rng(int64_t seed) {
 void destroy_rng(void) {
     if(r)gsl_rng_free(r);
 }
-static inline uint64_t RandomSuccessor(uint32_t v,uint32_t coord,double *weights,uint32_t *adj_list) {
+static inline uint64_t RandomSuccessor(uint64_t v,uint64_t coord,double *weights,uint64_t *adj_list) {
     if(!weights) return 0;
     double urv =gsl_rng_uniform(r);
-    uint32_t i,offset=v*coord;
+    uint64_t i,offset=v*coord;
     for(i=0;i<coord;++i)
         if(weights[offset+i] > urv) return adj_list[offset+i]; 
     return 3;
 } 
 
 
-uint64_t RandomTreeWithRoot(uint32_t r,uint32_t nv,uint32_t coord, void *_InTree, 
+uint64_t RandomTreeWithRoot(uint64_t r,uint64_t nv,uint64_t coord, void *_InTree, 
         void *_adj_list, void *_Next, void *_weights) {
     
     if(!_adj_list || !_InTree || !_Next || !_weights || !r) return 0;
     uint8_t * InTree = (uint8_t *)_InTree;
-    uint32_t * adj_list = (uint32_t *)_adj_list;
-    uint32_t * Next = (uint32_t *)_Next;
+    uint64_t * adj_list = (uint64_t *)_adj_list;
+    uint64_t * Next = (uint64_t *)_Next;
     double * weights = (double *)_weights;
 
     if(r >= nv) return 0;
-    uint32_t iterations=0;
+    uint64_t iterations=0;
     InTree = memset(InTree,0,nv);
-    uint32_t i=0,u;
+    uint64_t i=0,u;
     Next[r] = -1;
     InTree[r] = 1;
     for(;i<nv;++i) {
